Working Juicer script for Hi-C Chrom. Contact data
Read me is on file header. The original Juicer.sh script is a complex, large script that is dependent in my opinion on a lot of default in-house programs and old versions found in the HPC the script was developed in. I have modified the script, adapted and optimised it to run in Auburn University's "Easley" HPC. Easley, as well as many other HPCs are rather restrictive (by cybersecurity standards) to not allow scripts to run amok calling functions and creating jobs that the user has not specified. This means that the original Juicer script cannot completely finish the job on its own. This script therefore includes an extra custom script and some usage information at the header to indicate at which point the user has to take over the job to finish it manually.